requirements.txt
flask
mysql-connector-python
azure-identity
azure-keyvault-secrets
pyyaml
flasgger